use untitled file to run tests

+++ clients/cobol-lsp-vscode-extension/src/test/runTest.ts
@@ -31,11 +31,13 @@ async function main() {
     const userDir = os.tmpdir();
     if (process.argv.indexOf("--native") != -1) {
       console.log("Running tests with native language server");
-      fs.mkdirSync(path.join(userDir, "User"));
-      fs.writeFileSync(
-        path.join(userDir, "User", "settings.json"),
-        '{"cobol-lsp.serverRuntime": "NATIVE"}',
-      );
+      if (!fs.existsSync(path.join(userDir, "User"))) {
+        fs.mkdirSync(path.join(userDir, "User"));
+        fs.writeFileSync(
+          path.join(userDir, "User", "settings.json"),
+          '{"cobol-lsp.serverRuntime": "NATIVE"}',
+        );
+      }
     }
 
     const launchArgs = [
